chore(sdk-core): refactor and rename nTilde -> ntilde
Ticket: BG-72143
